refactor(editor): Move the shell modal catalogue to stores/defaults/modals.ts (no-changelog)
